Remove !mcd and move event listeners to new file
