+++ WiertarBot/unlock.py
@@ -1,5 +1,4 @@
 from selenium import webdriver
-from webdriver_manager.chrome import ChromeDriverManager
 
 import time
 import imaplib
