+++ main.py
@@ -2,8 +2,6 @@ import asyncio
 
 from WiertarBot import WiertarBot
 
-loop = asyncio.get_event_loop()
 
-bot = WiertarBot(loop)
-
-loop.run_forever()
+bot = WiertarBot()
+asyncio.get_event_loop().run_forever()
