+++ requirements.txt
@@ -1,7 +1,6 @@
 aiofiles==0.7.0
 aiohttp==3.7.4.post0
 async-timeout==3.0.1
-baseconvert==1.0.0a4
 bs4==0.0.1
 beautifulsoup4==4.10.0
 fbchat-asyncio==0.6.21
@@ -13,4 +12,3 @@ soupsieve==2.2.1
 aiogoogletrans==3.3.2
 cloudscraper==1.2.58
 selenium==3.141.0
-webdriver-manager==3.4.2
